Vignette added Contrast control added Bug fixes Added bar numb control, bar size control and more and more... Palette bar position control added

+++ app/src/main/java/net/henryco/opalette/api/glES/render/graphics/fbo/FrameBuffer.java
@@ -228,8 +228,8 @@ public class FrameBuffer implements OPallFBO {
 		}
 	}
 
-	public FrameBuffer createFBO(int w, int h, boolean depth) {
-		return createFBO(w, h, w, h, depth);
+	public FrameBuffer createFBO(float w, float h, boolean depth) {
+		return createFBO((int)w, (int)h, (int)w, (int)h, depth);
 	}
 
 	@Override
